Split test_wrapper_tensor.py to test_wrapper_tensor.py and test_wrapper_ops.py
